+++ pyccel/ast/cwrapper.py
@@ -718,7 +718,8 @@ class PyClassDef(ClassDef):
         wrapped.
     """
     __slots__ = ('_original_class', '_struct_name', '_type_name', '_type_object',
-                 '_new_func', '_properties')
+                 '_new_func', '_properties', '_number_magic_methods')
+    _attribute_nodes = ClassDef._attribute_nodes + ('_number_magic_methods',)
 
     def __init__(self, original_class, struct_name, type_name, scope, **kwargs):
         self._original_class = original_class
@@ -727,6 +728,7 @@ class PyClassDef(ClassDef):
         self._type_object = Variable(PyccelPyClassType(), type_name)
         self._new_func = None
         self._properties = ()
+        self._number_magic_methods = ()
         variables = [Variable(VoidType(), 'instance', memory_handling='alias'),
                      Variable(PyccelPyObject(), 'referenced_objects', memory_handling='alias'),
                      Variable(PythonNativeBool(), 'is_alias')]
@@ -818,6 +820,33 @@ class PyClassDef(ClassDef):
         Get all wrapped class properties.
         """
         return self._properties
+
+    def add_new_magic_number_method(self, method):
+        """
+        Add a new magic number method to the current class.
+
+        Add a new magic method to the current ClassDef describing
+        a number method.
+
+        Parameters
+        ----------
+        method : FunctionDef
+            The Method that will be added.
+        """
+
+        if not isinstance(method, PyFunctionDef):
+            raise TypeError("Method must be FunctionDef")
+        method.set_current_user_node(self)
+        self._number_magic_methods += (method,)
+
+    @property
+    def number_magic_methods(self):
+        """
+        Get the magic methods describing number methods.
+
+        Get the magic methods describing number methods such as __add__.
+        """
+        return self._number_magic_methods
 
 #-------------------------------------------------------------------