+++ pyccel/codegen/printing/fcode.py
@@ -26,6 +26,8 @@ from pyccel.ast.builtins import PythonBool, PythonList, PythonSet, VariableItera
 
 from pyccel.ast.builtin_methods.dict_methods import DictItems
 
+from pyccel.ast.builtin_methods.list_methods import ListPop
+
 from pyccel.ast.builtin_methods.set_methods import SetUnion
 
 from pyccel.ast.core import FunctionDef, FunctionDefArgument, FunctionDefResult
@@ -69,7 +71,7 @@ from pyccel.ast.numpyext import NumpyNonZero
 from pyccel.ast.numpyext import NumpySign
 from pyccel.ast.numpyext import NumpyIsFinite, NumpyIsNan
 
-from pyccel.ast.numpytypes import NumpyNDArrayType
+from pyccel.ast.numpytypes import NumpyNDArrayType, NumpyInt64Type
 
 from pyccel.ast.operators import PyccelAdd, PyccelMul, PyccelMinus, PyccelAnd, PyccelEq
 from pyccel.ast.operators import PyccelMod, PyccelNot, PyccelAssociativeParenthesis
@@ -1196,6 +1198,25 @@ class FCodePrinter(CodePrinter):
         arg_code = self._get_node_without_gFTL(expr.arg)
         return f"abs({arg_code})"
 
+    def _print_PythonRound(self, expr):
+        arg = expr.arg
+        if not isinstance(arg.dtype.primitive_type, PrimitiveFloatingPointType):
+            arg = self._apply_cast(NumpyInt64Type(), arg)
+        self.add_import(Import('pyc_math_f90', Module('pyc_math_f90',(),())))
+
+        arg_code = self._print(arg)
+        ndigits = self._apply_cast(NumpyInt64Type(), expr.ndigits) if expr.ndigits \
+                else LiteralInteger(0, NumpyInt64Type())
+        ndigits_code = self._print(ndigits)
+
+        code = f'pyc_bankers_round({arg_code}, {ndigits_code})'
+
+        if not isinstance(expr.dtype.primitive_type, PrimitiveFloatingPointType):
+            prec = self.print_kind(expr)
+            return f"Int({code}, kind={prec})"
+        else:
+            return code
+
     def _print_PythonTuple(self, expr):
         shape = tuple(reversed(expr.shape))
         if len(shape)>1:
@@ -1310,9 +1331,42 @@ class FCodePrinter(CodePrinter):
     #========================== List Methods ===============================#
 
     def _print_ListAppend(self, expr):
-        target = expr.list_obj
+        target = self._print(expr.list_obj)
         arg = self._print(expr.args[0])
         return f'call {target} % push_back({arg})\n'
+
+    def _print_ListPop(self, expr):
+        list_obj = expr.list_obj
+        target = self._print(list_obj)
+        index_element = expr.index_element
+        parent_assign_nodes = expr.get_direct_user_nodes(lambda u: isinstance(u, Assign))
+        if parent_assign_nodes:
+            lhs = expr.current_user_node.lhs
+        else:
+            lhs = self.scope.get_temporary_variable(expr.class_type)
+
+        lhs_code = self._print(lhs)
+
+        if index_element:
+            _shape = PyccelArrayShapeElement(list_obj, index_element)
+            if isinstance(index_element, PyccelUnarySub) and isinstance(index_element.args[0], LiteralInteger):
+                index_element = PyccelMinus(_shape, index_element.args[0], simplify = True)
+            tmp_iter = self.scope.get_temporary_variable(IteratorType(list_obj.class_type),
+                    name = f'{list_obj}_iter')
+            code = (f'{tmp_iter} = {target} % begin() + {self._print(index_element)}\n'
+                    f'{lhs} = {tmp_iter} % of()\n'
+                    f'{tmp_iter} = {target} % erase({tmp_iter})\n')
+        else:
+            index = expr.index_element or PyccelUnarySub(LiteralInteger(1))
+            rhs = self._print(IndexedElement(list_obj, index))
+            code = (f'{lhs_code} = {rhs}\n'
+                    f'call {target} % pop_back()\n')
+
+        if parent_assign_nodes:
+            return code
+        else:
+            self._additional_code += code
+            return lhs_code
 
     #========================== Set Methods ================================#
 
@@ -2017,7 +2071,7 @@ class FCodePrinter(CodePrinter):
     def _print_Assign(self, expr):
         rhs = expr.rhs
 
-        if isinstance(rhs, (FunctionCall, SetUnion)):
+        if isinstance(rhs, (FunctionCall, SetUnion, ListPop)):
             return self._print(rhs)
 
         lhs_code = self._print(expr.lhs)